+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # a-type-of-js
 
-## 0.0.0 (3 月 4 2025 年)
+## 0.0.1 (3 月 7 日 2025 年)
+
+- 移除了 `isAtomics` 方法。因为 `Atomics` 仅是一个方法集合，没有实例。同样的还有 `Math、JSON、Reflect、Intl`
+- 移除了 `isWindow` 方法
+- 优化了 `typeOf` 方法，检测 `asyncFunction`、`generator`、`generatorFunction` 更精准
+
+## 0.0.0 (3 月 4 日 2025 年)
 
 - 初始化项目